Support shifting items to different stages during migration
Support shifting items to a child, parent or parallel stage

If a PlanItem is moved to a Stage that is not yet Active, then the item is in "moving" state.
This is captured with the PlanItemMoving event.
If the new stage of the item is already Active, then the event PlanItemMoved is also invoked.
Alternatively, in case of Available or Suspended stages, the PlanItemMoved event is added when the stage becomes Active again.
PlanItemMoving disconnects the PlanItem from the SentryNetwork.
PlanItemMoved reconnects the PlanItem from the SentryNetwork.

Note: This change does not cover situations where plan items shift to stages that are repeating and for which multiple instances are active.
Current migration support in general does not deal with specific repetition patterns.